In the previous fitness function, if on a cloud the first app is a task, and all others are services, all services will have very low fitness, because "the start time of all services are equal to the task completion time on this cloud", so the previous fitness function is not reasonable.
In our new fitness function, we only consider the time of every app and whether it is accepted. We set a linear function from app time to the fitness. If an app is rejected, we set it as a very big value, currently 1800s, because if a task or a service runs too late, it is close to being rejected.

I do not need to consider the ratio of the fitness values, because the standardization of my fitness function can give them a good scale.